Topics/Python/02-cod3r/trabalho-de-conclusao/backend/python/dev/requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
